ImageResizer/requirements.txt
Pillow==9.4.0
pika==1.3.1